+++ lib/fiber/array_4d.hpp
@@ -34,16 +34,16 @@ class Array4d
 {
 public:
     Array4d();
-    Array4d(int extent0, int extent1, int extent2, int extent3);
-    Array4d(int extent0, int extent1, int extent2, int extent3, T* data);
+    Array4d(size_t extent0, size_t extent1, size_t extent2, size_t extent3);
+    Array4d(size_t extent0, size_t extent1, size_t extent2, size_t extent3, T* data);
 
     ~Array4d();
 
-    T& operator()(int index0, int index1, int index2, int index3);
-    const T& operator()(int index0, int index1, int index2, int index3) const;
+    T& operator()(size_t index0, size_t index1, size_t index2, size_t index3);
+    const T& operator()(size_t index0, size_t index1, size_t index2, size_t index3) const;
 
-    int extent(int dimension) const;
-    void set_size(int extent0, int extent1, int extent2, int extent3);
+    size_t extent(size_t dimension) const;
+    void set_size(size_t extent0, size_t extent1, size_t extent2, size_t extent3);
 
     typedef T* iterator;
     typedef const T* const_iterator;
@@ -55,9 +55,9 @@ public:
 
 private:
 #ifdef FIBER_CHECK_ARRAY_BOUNDS
-    void check_dimension(int dimension) const;
-    void check_extents(int extent0, int extent1, int extent2, int extent3) const;
-    void check_indices(int index0, int index1, int index2, int index3) const;
+    void check_dimension(size_t dimension) const;
+    void check_extents(size_t extent0, size_t extent1, size_t extent2, size_t extent3) const;
+    void check_indices(size_t index0, size_t index1, size_t index2, size_t index3) const;
 #endif
 
 private:
@@ -66,7 +66,7 @@ private:
     Array4d& operator=(const Array4d& rhs);
 
 private:
-    int m_extents[4];
+    size_t m_extents[4];
     bool m_owns;
     T* m_storage;
 };
