+++ lib/fiber/function.hpp
@@ -39,10 +39,10 @@ public:
 
     virtual ~Function() {}
 
-    virtual int worldDimension() const = 0;
-    virtual int codomainDimension() const = 0;
+    virtual size_t worldDimension() const = 0;
+    virtual size_t codomainDimension() const = 0;
 
-    virtual void addGeometricalDependencies(int& geomDeps) const = 0;
+    virtual void addGeometricalDependencies(size_t& geomDeps) const = 0;
 
     virtual void evaluate(const GeometricalData<CoordinateType>& geomData,
                           arma::Mat<ValueType>& result) const = 0;
